refactor: extract ReadingPositionService into ReadingSessionManager
